+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     * Added optional password/database attributes for redis in `system.yaml`
     * Added ability to filter enabled or disabled with bin/gpm index [#3187](https://github.com/getgrav/grav/pull/3187)
     * Added `$grav->getVersion()` or `grav.version` in twig to get the current Grav version [#3142](https://github.com/getgrav/grav/issues/3142)
+    * Added second parameter to `$blueprint->flattenData()` to include every field, including those which have no data
 1. [](#bugfix)
     * Fixed CLI progressbar in `backup` and `security` commands to use styled output [#3198](https://github.com/getgrav/grav/issues/3198)
     * Fixed page save failing because of uploaded images [#3191](https://github.com/getgrav/grav/issues/3191)